% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endDiv.R
\name{endDiv}
\alias{endDiv}
\title{Subpixel spectral diversity}
\usage{
endDiv(x)
}
\arguments{
\item{x}{RasterStack, RasterBrick.  Each layer displays the abundance map of an endmember.}
}
\value{
Endmember diversity per pixel as RasterBrick. The four layers include Richness, Shannon-Weaver index, Simpson index and Evenness
calculated based on the endmember abundance map
}
\description{
Calculates different diversity indices for each pixel based on the endmember abundance map.
}
\references{
Rossi, C., and Gholizadeh, H. (2022). Subpixel spectral diversity: Using endmember diversity to capture plant diversity.
submitted to Methods in Ecology and Evolution.
}
\author{
Christian Rossi christian.rossi1990@gmail.com
}
