% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfcvd.R
\name{hfcvd}
\alias{hfcvd}
\title{Noise-whitened Harsanyi–Farrand–Chang method}
\usage{
hfcvd(x,noisewithen, far)
}
\arguments{
\item{x}{RasterStack, matrix with dimensions L(bands) x N(pixels). Remote sensing imagery (usually hyperspectral),}

\item{noisewithen}{Logical. If TRUE, performs noise-whitening before computing the number of endmembers.
FALSE consist in the Harsanyi–Farrand–Chang (HFC) method. By default TRUE.}

\item{far}{Positive integer.  Probability of false alarm . By default 10^-3}
}
\value{
Number of endmembers in the image, returned as a positive integer.
}
\description{
Finds the number of endmembers present in a multispectral or hyperspectral image by using the noise-whitened Harsanyi–Farrand–Chang (NWHFC) method.
}
\references{
C.-I. Chang and Q. Du, "Estimation of number of spectrally distinct
signal sources in hyperspectral imagery," IEEE Transactions on Geoscience and Remote Sensing, vol. 43, no. 3, mar 2004.
}
\author{
Christian Rossi christian.rossi1990@gmail.com
}
