% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysime.R
\name{hysime}
\alias{hysime}
\title{HySime}
\usage{
hysime(x, noisetype)
}
\arguments{
\item{x}{RasterStack, matrix with dimensions L(bands) x N(pixels). Remote sensing imagery (usually hyperspectral).}

\item{noisetype}{Character. Infers the noise in a image, by assuming that the reflectance at a given band is well modelled by
a linear regression on the remaining bands. By default the noise type is "additive". Other option is "poisson".}
}
\value{
Number of endmembers or virtual dimensionality of the image, returned as a positive integer.
}
\description{
Hyperspectral signal subspace estimation to find the number of endmembers present in a multispectral or hyperspectral image
}
\references{
J. M. Bioucas-Dias and J. M. P. Nascimento, "Hyperspectral Subspace Identification,"
in IEEE Transactions on Geoscience and Remote Sensing, vol. 46, no. 8, pp. 2435-2445, Aug. 2008.
}
\author{
Christian Rossi christian.rossi1990@gmail.com
}
