% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\name{ppi}
\alias{ppi}
\title{Pixel purity index}
\usage{
ppi(x, q, numSkewers, reductionmethod)
}
\arguments{
\item{x}{RasterStack, matrix with dimensions L(bands) x N(pixels). Remote sensing imagery (usually hyperspectral),}

\item{q}{Positive integer. number of endmembers in the scene}

\item{numSkewers}{Positive integer. Number of randomly generated unit vectors to be used for projection. By default 10^4}

\item{reductionmethod}{Character. Select a dimensionality reduction method. By default maximum noise fraction "MNF". Other options are
principal component analysis "PCA" or "None".}
}
\value{
Estimated mixing matrix (endmembers signatures)
}
\description{
Extracts endmember signatures from multispectral or hyperspectral image by using the pixel purity index (PPI) algorithm.
}
\references{
J.W Boardman, F.A. Kruse and R.O. Green, "Mapping target signatures via partial unmixing of AVIRIS data.",
Technical Report, California, USA, 1995.
}
\author{
Christian Rossi christian.rossi1990@gmail.com
}
