% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vca.R
\name{vca}
\alias{vca}
\title{Vertex Component Analysis}
\usage{
vca(x, q)
}
\arguments{
\item{x}{RasterStack, matrix with dimensions L(bands) x N(pixels). Remote sensing imagery (usually hyperspectral).}

\item{q}{Positive integer. number of endmembers in the scene}
}
\value{
Estimated mixing matrix (endmembers signatures)
}
\description{
Extracts endmember signatures from multispectral or hyperspectral image by using the Vertex Component Analysis method (VCA).
}
\references{
José M. P. Nascimento and José M. B. Dias "Vertex Component Analysis: A Fast Algorithm to Unmix Hyperspectral Data"
submited to IEEE Trans. Geosci. Remote Sensing, vol. .., no. .., pp. .-., 2004
}
\author{
Christian Rossi christian.rossi1990@gmail.com
}
